redact page
редакт и по мелочи

+++ src/app/cards/pages/card-page/card-page.component.ts
@@ -1,42 +1,44 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { Card } from 'src/app/app.component';
-import { GetCardService } from 'src/app/services/get-card.service';
+import { ICard } from 'src/app/app.component';
+import { CardService } from 'src/app/services/card.service';
 
 @Component({
-  selector: 'app-card-page',
-  templateUrl: './card-page.component.html',
-  styleUrls: ['./card-page.component.scss'],
+    selector: 'app-card-page',
+    templateUrl: './card-page.component.html',
+    styleUrls: ['./card-page.component.scss'],
 })
 export class CardPageComponent implements OnInit {
-  card!: Card;
+    public card!: ICard;
 
-  constructor(
-    private _route: ActivatedRoute,
-    private _router: Router,
-    private _cardService: GetCardService,
-  ) { }
+    constructor(
+        private _route: ActivatedRoute,
+        private _router: Router,
+        private _cardService: CardService,
+    ) { }
 
-  ngOnInit(): void {
-    this._route.params.subscribe((params: Params) => {
-      this.card = this._cardService.getCardById(+params['id']);
-    });
-  }
+    public ngOnInit(): void {
+        this._route.params.subscribe((params: Params) => {
+            this.card = this._cardService.getCardById(+params['id']);
+        });
+    }
 
-  onClickBack() {
-    this._router.navigate(['main-page']);
-  }
+    public onClickBack(): void {
+        this._router.navigate(['main-page']);
+    }
 
-  onDeleteCard() {
-    this._route.params.subscribe((params: Params) => {
-      this._cardService.deleteCardById(+params['id']);
-    });
-    this._router.navigate(['main-page']);
-  }
+    public async onDeleteCard(): Promise<void> {
+        if (await this._cardService.getConfirm(`УДАЛИТЬ КАРТУ?`)) {
+            this._route.params.subscribe((params: Params) => {
+                this._cardService.deleteCardById(+params['id']);
+            });
+            this._router.navigate(['main-page']);
+        }
+    }
 
-  onRedactCard() {
-    this._route.params.subscribe((params: Params) => {
-      this._router.navigate(['redact'], { relativeTo: this._route })
-    });
-  }
+    public onRedactCard(): void {
+        this._route.params.subscribe((params: Params) => {
+            this._router.navigate(['redact'], { relativeTo: this._route });
+        });
+    }
 }
